Don't import test files inside production code paths
