+++ models/demos/deepseek_v3_b1/tests/unit_tests/test_decoder_block.py
@@ -21,12 +21,12 @@ from models.demos.deepseek_v3_b1.fused_ops.attention_block.op import AttentionBl
 from models.demos.deepseek_v3_b1.fused_ops.decoder_block.op import DecoderBlock
 from models.demos.deepseek_v3_b1.fused_ops.moe.op import MoeOp
 from models.demos.deepseek_v3_b1.micro_ops.flash_mla.op import FlashMLADecode
+from models.demos.deepseek_v3_b1.model_dimensions import RoutedExpert
 from models.demos.deepseek_v3_b1.tests.unit_tests.ccl_test_utils import create_fabric_router_config
 from models.demos.deepseek_v3_b1.tests.unit_tests.test_moe_mlp import (
     DENSE_LAYER_IDX,
     DENSE_SHARED_N,
     ROUTED_EXPERT_LAYER_IDX,
-    RoutedExpert,
     extract_routed_expert_output,
 )
 from models.demos.deepseek_v3_b1.weights.prepare import get_layer_raw_tensors
